project folder restructure to match what we discussed
