--- src/components/sellerPages/profile/dashboard/invSection/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import SmallItemCard from '../../../../common/cards/smallItem';
-
-function InvSection() {
-  return (
-    <>
-      <h2>Inventory</h2>
-      <h4>Details</h4>
-      <SmallItemCard headerText="Option Text" descText="Subtext" />
-    </>
-  );
-}
-export default InvSection;
