+++ src/pages/InventoryPage/inventoryStyles.css
