+++ src/pages/SellerProfilePage/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { NavBar } from '../../components/common/NavBar';
+import Dashboard from '../../components/dashboard/dashboard';
+
+function SellerProfile() {
+  return (
+    <>
+      <NavBar />
+      <Dashboard />
+    </>
+  );
+}
+
+export default SellerProfile;
