+++ src/state/actions/index.js
@@ -5,7 +5,7 @@ import {
   getProfileData,
   getDSData,
   postData,
-} from '../../api/index';
+} from '../../services/api/index';
 
 export const FETCH_PRODUCTS_START = 'FETCH_PRODUCTS_START';
 export const FETCH_PRODUCTS_SUCCESS = 'FETCH_PRODUCTS_SUCCESS';
